Refactor MCP DTO parser module
